+++ app.js
@@ -2,7 +2,7 @@ const express = require('express')
 const logger = require('morgan')
 const cors = require('cors')
 
-const contactsRouter = require('./routes/api/contacts');
+const routers =  require('./routes/api/contacts');
 
 const app = express()
 
@@ -12,7 +12,8 @@ app.use(logger(formatsLogger))
 app.use(cors())
 app.use(express.json())
 
-app.use('/api/contacts', contactsRouter)
+app.use('/api/contacts', routers)
+
 
 app.use((req, res) => {
   res.status(404).json({ message: 'Not found' })
@@ -22,5 +23,5 @@ app.use((err, req, res, next) => {
   const { status = 500, message = 'Server error' } = err;
   res.status(status).json({ message })
 })
-
 module.exports = app
+
