+++ controllers/contacts/index.js
@@ -1,0 +1,13 @@
+const listContacts = require('./listContacts');
+const getContactById = require('./getContactById');
+const addContact = require('./addContact');
+const updateContactById = require('./updateContactById');
+const removeContact = require('./removeContact');
+
+module.exports = {
+  listContacts,
+  getContactById,
+  removeContact,
+  addContact,
+  updateContactById,
+}
