+++ controllers/index.js
@@ -1,13 +1,5 @@
-const listContacts = require('./listContacts');
-const getContactById = require('./getContactById');
-const addContact = require('./addContact');
-const updateContactById = require('./updateContactById');
-const removeContact = require('./removeContact');
+const contacts = require("./contacts");
 
 module.exports = {
-  listContacts,
-  getContactById,
-  removeContact,
-  addContact,
-  updateContactById,
+    contacts
 }
