+++ middlewares/index.js
@@ -1,3 +1,4 @@
 const validation = require('./validation');
 
 module.exports = validation;
+
