+++ routes/api/contacts.js
@@ -1,23 +1,21 @@
 const express = require('express')
 
-const {contacts:ctrl} = require('../../controllers')
-
+const {contacts} = require("../../controllers");
 const { ctrlWrapper } = require('../../helpers')
 
-const { validation } = require('../../middlewares')
-
+const  validation  = require('../../middlewares')
 const { schemaAdd, schemaUpdate } = require('../../schemas/contacts');
 
 const router = express.Router();
 
-router.get('/', ctrlWrapper(ctrl.listContacts))
+router.get('/', ctrlWrapper(contacts.listContacts))
 
-router.get('/:id', ctrlWrapper(ctrl.getContactById))
+router.get('/:id', ctrlWrapper(contacts.getContactById))
 
-router.post('/', validation(schemaAdd), ctrlWrapper(ctrl.addContact));
+router.post('/', validation(schemaAdd), ctrlWrapper(contacts.addContact));
 
-router.delete('/:id',ctrlWrapper(ctrl.removeContact))
+router.delete('/:id',ctrlWrapper(contacts.removeContact))
 
-router.put('/:id',validation(schemaUpdate), ctrlWrapper(ctrl.updateContact))
+router.put('/:id',validation(schemaUpdate), ctrlWrapper(contacts.updateContact))
 
 module.exports = router;
